[funexpected.export] remove extra files from exported pack (#25)
* [funexpected.export] remove extra files from exported pack

* [funexpected.export] remove project.binary from exported pack

+++ editor/editor_export.cpp
@@ -235,10 +235,21 @@ String EditorExportPreset::get_script_encryption_key() const {
 	return script_key;
 }
 
+bool EditorExportPreset::get_export_non_resource_files() const {
+
+	return export_non_resource_files;
+}
+
+void EditorExportPreset::set_export_non_resource_files(bool mode) {
+
+	export_non_resource_files = mode;
+}
+
 EditorExportPreset::EditorExportPreset() :
 		export_filter(EXPORT_ALL_RESOURCES),
 		export_path(""),
 		runnable(false),
+		export_non_resource_files(true),
 		script_mode(MODE_SCRIPT_COMPILED) {
 }
 
@@ -932,25 +943,27 @@ Error EditorExportPlatform::export_project_files(const Ref<EditorExportPreset> &
 		}
 	}
 
-	// Store icon and splash images directly, they need to bypass the import system and be loaded as images
-	String icon = ProjectSettings::get_singleton()->get("application/config/icon");
-	String splash = ProjectSettings::get_singleton()->get("application/boot_splash/image");
-	if (icon != String() && FileAccess::exists(icon)) {
-		Vector<uint8_t> array = FileAccess::get_file_as_array(icon);
-		p_func(p_udata, icon, array, idx, total);
-	}
-	if (splash != String() && FileAccess::exists(splash) && icon != splash) {
-		Vector<uint8_t> array = FileAccess::get_file_as_array(splash);
-		p_func(p_udata, splash, array, idx, total);
-	}
+	if (p_preset->get_export_non_resource_files()) {
+		// Store icon and splash images directly, they need to bypass the import system and be loaded as images
+		String icon = ProjectSettings::get_singleton()->get("application/config/icon");
+		String splash = ProjectSettings::get_singleton()->get("application/boot_splash/image");
+		if (icon != String() && FileAccess::exists(icon)) {
+			Vector<uint8_t> array = FileAccess::get_file_as_array(icon);
+			p_func(p_udata, icon, array, idx, total);
+		}
+		if (splash != String() && FileAccess::exists(splash) && icon != splash) {
+			Vector<uint8_t> array = FileAccess::get_file_as_array(splash);
+			p_func(p_udata, splash, array, idx, total);
+		}
 
-	String config_file = "project.binary";
-	String engine_cfb = EditorSettings::get_singleton()->get_cache_dir().plus_file("tmp" + config_file);
-	ProjectSettings::get_singleton()->save_custom(engine_cfb, custom_map, custom_list);
-	Vector<uint8_t> data = FileAccess::get_file_as_array(engine_cfb);
-	DirAccess::remove_file_or_error(engine_cfb);
+		String config_file = "project.binary";
+		String engine_cfb = EditorSettings::get_singleton()->get_cache_dir().plus_file("tmp" + config_file);
+		ProjectSettings::get_singleton()->save_custom(engine_cfb, custom_map, custom_list);
+		Vector<uint8_t> data = FileAccess::get_file_as_array(engine_cfb);
+		DirAccess::remove_file_or_error(engine_cfb);
 
-	p_func(p_udata, "res://" + config_file, data, idx, total);
+		p_func(p_udata, "res://" + config_file, data, idx, total);
+	}
 
 	return OK;
 }
@@ -1247,6 +1260,8 @@ void EditorExport::_save() {
 		config->set_value(section, "export_path", preset->get_export_path());
 		config->set_value(section, "script_export_mode", preset->get_script_export_mode());
 		config->set_value(section, "script_encryption_key", preset->get_script_encryption_key());
+		if (!preset->get_export_non_resource_files())
+			config->set_value(section, "export_non_resource_files", preset->get_export_non_resource_files());
 
 		String option_section = "preset." + itos(i) + ".options";
 
@@ -1472,7 +1487,10 @@ void EditorExport::load_config() {
 		if (config->has_section_key(section, "script_encryption_key")) {
 			preset->set_script_encryption_key(config->get_value(section, "script_encryption_key"));
 		}
-
+		if (config->has_section_key(section, "export_non_resource_files")) {
+			preset->set_export_non_resource_files(config->get_value(section, "export_non_resource_files"));
+		}
+		
 		String option_section = "preset." + itos(index) + ".options";
 
 		List<String> options;
